quoting of values containing colons ':'
A ':' as last or single character of a value
confuses YAML readers. The most consistent
behaviour among all YAML parsers (i.e, YAML,
YAML::XS, Data::YAML::Reader) is when we
quote all values that end with ':'.

We could quote generally everything that contains
a quote but hereby we only fix the minimum issue
which definitely did not work so far, so we don't
risk introducing backwards incompatibilities.