Fix YAMLish behaviour with empty values
Same fix as proposed for TAP::Parser::YAMLish::Reader.

In a hash with non-existent values, like this:

  ok - foo
    ---
    foo:
    bar: 2
    baz: 3
    ...

the parser now returns

  {
   foo => undef,
   bar => 2,
   baz => 3,
  }

instead of wrongly assuming a new hash would start
a new sub-structure, like

  {
   foo => {
           bar => 2,
           baz => 3,
          }
  }

which went totally mental if you have a dozen of
such key:value lines with some of them missing a
value, which led to a deeply nested structure that
has nothing in common with the structure it returns
when the keys exist.

Now it is consistently parsing the value of 'foo'
as scalar, and provides undef for missing values.

That behaviour is also consistent with the most
common YAML modules on CPAN (YAML::Tiny, YAML::Syck,
YAML::Old, YAML::XS, YAML::PP).

The only dissent in above modules is whether a missing
value gets undef or empty string ('').

The YAML 1.2 spec explicitely specifies to resolve such
"completely empty nodes" as "null value" (see
chapter 7.2. Empty Nodes). The YAML 1.1 spec is not that
explicit but also defines a special glyph to describe
such "completely empty values". So using undef in Perl
sounds most spec-conforming and is also what the modules
YAML::{XS,Old,Syck) do.